+++ src/client/client.cpp
@@ -1,7 +1,6 @@
-#include "session.hpp"
+#include "client.hpp"
 #include "game/constants.hpp"
 #include "game/entity_type.hpp"
-#include "net/client.hpp"
 #include "serial/serialize.hpp"
 
 #include <Eigen/Geometry>
@@ -17,15 +16,13 @@ Eigen::Quaternionf yaw_orientation(float yaw) {
 }
 } // namespace
 
-Session::Session(Session_create_info const &info)
-    : net::Client{{
-        .incoming_bandwidth = info.incoming_bandwidth,
-        .outgoing_bandwidth = info.outgoing_bandwidth,
-      }},
-      _tick_duration{info.tick_duration} {}
+Client::Client(Client_create_info const &info)
+    : net::Client{info.net_info}, _tick_duration{info.tick_duration} {}
 
-void Session::service_input_tick(float duration) {
-  assert(has_scene());
+void Client::tick(float duration) {
+  if (!has_scene()) {
+    return;
+  }
   _tick_timer -= duration;
   if (_tick_timer <= 0) {
     auto const tracy_frame_name = "Input Tick";
@@ -33,8 +30,7 @@ void Session::service_input_tick(float duration) {
     try {
       _tick_timer += _tick_duration;
       if (_player_entity_id) {
-        net::Client::send_player_input_state(
-          *_player_entity_id, _current_input_state);
+        send_player_input_state(*_player_entity_id, _current_input_state);
       }
     } catch (...) {
       FrameMarkEnd(tracy_frame_name);
@@ -44,56 +40,35 @@ void Session::service_input_tick(float duration) {
   }
 }
 
-bool Session::has_scene() const noexcept { return _scene.has_value(); }
+bool Client::has_scene() const noexcept { return _scene.has_value(); }
 
-void Session::poll_events() {
-  ZoneScoped;
-  net::Client::poll_events();
-}
-
-void Session::wait_events(std::uint32_t timeout) {
-  net::Client::wait_events(timeout);
-}
-
-void Session::connect(enet::Address const &address) {
-  net::Client::connect(address);
-}
-
-bool Session::is_connecting() const noexcept {
-  return net::Client::is_connecting();
-}
-
-bool Session::is_connected() const noexcept {
-  return net::Client::is_connected();
-}
-
-net::Input_state const &Session::get_current_input_state() const noexcept {
+net::Input_state const &Client::get_current_input_state() const noexcept {
   return _current_input_state;
 }
 
-void Session::set_current_input_state(net::Input_state const &value) noexcept {
+void Client::set_current_input_state(net::Input_state const &value) noexcept {
   _current_input_state = value;
 }
 
-scene::Scene *Session::get_scene() noexcept {
+Scene *Client::get_scene() noexcept {
   return _scene ? &*_scene : nullptr;
 }
 
-scene::Scene const *Session::get_scene() const noexcept {
+Scene const *Client::get_scene() const noexcept {
   return _scene ? &*_scene : nullptr;
 }
 
 std::optional<Camera_snapshot> const &
-Session::get_camera_snapshot() const noexcept {
+Client::get_camera_snapshot() const noexcept {
   return _camera_snapshot;
 }
 
-void Session::on_connect() {
-  _scene = scene::Scene{};
+void Client::on_connect() {
+  _scene = Scene{};
   send_player_join_request();
 }
 
-void Session::on_disconnect() {
+void Client::on_disconnect() {
   _tick_timer = 0.0f;
   _last_grid_state_payload.clear();
   _player_entity_id = std::nullopt;
@@ -102,14 +77,12 @@ void Session::on_disconnect() {
   _scene = std::nullopt;
 }
 
-void Session::on_player_join_response(net::Entity_id player_entity_id) {
+void Client::on_player_join_response(net::Entity_id player_entity_id) {
   _player_entity_id = player_entity_id;
   std::cout << "Got player join response. id = " << player_entity_id << ".\n";
 }
 
-void Session::on_update_grid() {}
-
-void Session::on_world_snapshot(
+void Client::on_world_snapshot(
   serial::Span_reader &grid_state_reader,
   serial::Span_reader &public_entity_state_reader,
   serial::Span_reader &player_entity_state_reader) {
@@ -121,14 +94,14 @@ void Session::on_world_snapshot(
     !std::ranges::equal(grid_state, _last_grid_state_payload);
   if (grid_state_changed) {
     _scene->get_grid().load(grid_state_reader);
+    _scene->set_grid_remesh_flag();
     _last_grid_state_payload.assign(grid_state.begin(), grid_state.end());
-    on_update_grid();
   }
   load_player_entity_state(player_entity_state_reader);
   load_public_entity_state(public_entity_state_reader);
 }
 
-void Session::load_player_entity_state(serial::Reader &reader) {
+void Client::load_player_entity_state(serial::Reader &reader) {
   using serial::deserialize;
   _local_humanoid_entity_id = std::nullopt;
   auto const entity_count = deserialize<std::uint32_t>(reader);
@@ -152,7 +125,7 @@ void Session::load_player_entity_state(serial::Reader &reader) {
   }
 }
 
-void Session::load_public_entity_state(serial::Reader &reader) {
+void Client::load_public_entity_state(serial::Reader &reader) {
   using serial::deserialize;
   auto &mesh_instances = _scene->get_mesh_instances();
   mesh_instances.clear();
@@ -187,8 +160,8 @@ void Session::load_public_entity_state(serial::Reader &reader) {
           .pitch = input_state->pitch,
         };
       } else {
-        mesh_instances.emplace_back(scene::Mesh_instance{
-          .mesh = scene::Mesh::cube,
+        mesh_instances.emplace_back(Mesh_instance{
+          .mesh = Mesh::cube,
           .position = *position + Eigen::Vector3f::UnitY() * 0.9f,
           .orientation = yaw_orientation(input_state->yaw),
           .scale = {0.7f, 1.8f, 0.7f},
@@ -203,8 +176,8 @@ void Session::load_public_entity_state(serial::Reader &reader) {
         std::cerr << "Malformed projectile state.\n";
         return;
       }
-      mesh_instances.emplace_back(scene::Mesh_instance{
-        .mesh = scene::Mesh::cube,
+      mesh_instances.emplace_back(Mesh_instance{
+        .mesh = Mesh::cube,
         .position = *position,
         .orientation = Eigen::Quaternionf::Identity(),
         .scale = Eigen::Vector3f::Constant(0.25f),
