+++ src/client/client.hpp
@@ -1,19 +1,18 @@
-#ifndef FPSPARTY_CLIENT_SESSION_HPP
-#define FPSPARTY_CLIENT_SESSION_HPP
+#ifndef FPSPARTY_CLIENT_CLIENT_HPP
+#define FPSPARTY_CLIENT_CLIENT_HPP
 
+#include "client/scene.hpp"
 #include "net/client.hpp"
 #include "net/core/entity_id.hpp"
 #include "net/core/input_state.hpp"
-#include "scene/scene.hpp"
 #include <Eigen/Dense>
 #include <cstddef>
 #include <optional>
 #include <vector>
 
 namespace fpsparty::client {
-struct Session_create_info {
-  std::uint32_t incoming_bandwidth{};
-  std::uint32_t outgoing_bandwidth{};
+struct Client_create_info {
+  net::Client_create_info net_info;
   float tick_duration;
 };
 
@@ -23,35 +22,25 @@ struct Camera_snapshot {
   float pitch{};
 };
 
-class Session : net::Client {
+class Client : public net::Client {
 public:
-  explicit Session(Session_create_info const &info);
+  explicit Client(Client_create_info const &info);
 
-  Session(Session const &other) = delete;
+  Client(Client const &other) = delete;
 
-  Session &operator=(Session const &other) = delete;
+  Client &operator=(Client const &other) = delete;
 
-  void service_input_tick(float duration);
+  void tick(float duration);
 
   bool has_scene() const noexcept;
-
-  void poll_events();
-
-  void wait_events(std::uint32_t timeout);
-
-  void connect(enet::Address const &address);
-
-  bool is_connecting() const noexcept;
-
-  bool is_connected() const noexcept;
 
   net::Input_state const &get_current_input_state() const noexcept;
 
   void set_current_input_state(net::Input_state const &value) noexcept;
 
-  scene::Scene *get_scene() noexcept;
+  Scene *get_scene() noexcept;
 
-  scene::Scene const *get_scene() const noexcept;
+  Scene const *get_scene() const noexcept;
 
   std::optional<Camera_snapshot> const &get_camera_snapshot() const noexcept;
 
@@ -61,8 +50,6 @@ protected:
   void on_disconnect() override;
 
   void on_player_join_response(net::Entity_id player_entity_id) override;
-
-  virtual void on_update_grid();
 
   void on_world_snapshot(
     serial::Span_reader &grid_state_reader,
@@ -74,7 +61,7 @@ private:
 
   void load_public_entity_state(serial::Reader &reader);
 
-  std::optional<scene::Scene> _scene{};
+  std::optional<Scene> _scene{};
   std::optional<net::Entity_id> _player_entity_id{};
   std::optional<net::Entity_id> _local_humanoid_entity_id{};
   std::optional<Camera_snapshot> _camera_snapshot{};
