+++ src/client/scene.cpp
@@ -1,6 +1,6 @@
 #include "scene.hpp"
 
-namespace fpsparty::scene {
+namespace fpsparty::client {
 game::Grid const &Scene::get_grid() const noexcept { return _grid; }
 
 game::Grid &Scene::get_grid() noexcept { return _grid; }
@@ -13,4 +13,12 @@ Scene::get_mesh_instances() const noexcept {
 std::vector<Mesh_instance> &Scene::get_mesh_instances() noexcept {
   return _mesh_instances;
 }
-} // namespace fpsparty::scene
+
+bool Scene::get_grid_remesh_flag() const noexcept {
+  return _grid_remesh_flag;
+}
+
+void Scene::set_grid_remesh_flag() noexcept { _grid_remesh_flag = true; }
+
+void Scene::clear_grid_remesh_flag() noexcept { _grid_remesh_flag = false; }
+} // namespace fpsparty::client
