+++ src/client/scene.hpp
@@ -1,12 +1,12 @@
-#ifndef FPSPARTY_SCENE_SCENE_HPP
-#define FPSPARTY_SCENE_SCENE_HPP
+#ifndef FPSPARTY_CLIENT_SCENE_HPP
+#define FPSPARTY_CLIENT_SCENE_HPP
 
 #include "game/grid.hpp"
 #include <Eigen/Dense>
 #include <Eigen/Geometry>
 #include <vector>
 
-namespace fpsparty::scene {
+namespace fpsparty::client {
 enum class Mesh {
   cube,
 };
@@ -28,10 +28,17 @@ public:
 
   std::vector<Mesh_instance> &get_mesh_instances() noexcept;
 
+  bool get_grid_remesh_flag() const noexcept;
+
+  void set_grid_remesh_flag() noexcept;
+
+  void clear_grid_remesh_flag() noexcept;
+
 private:
   game::Grid _grid{{}};
   std::vector<Mesh_instance> _mesh_instances{};
+  bool _grid_remesh_flag{};
 };
-} // namespace fpsparty::scene
+} // namespace fpsparty::client
 
 #endif
